pyproject.toml
[tool.poetry]
name = "financial-transactions"
version = "0.1.0"
description = ""
authors = ["penkhasoveg <pen.egor2002@gmail.com>"]
readme = "README.md"
packages = [{ include = "server", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.2"
djangorestframework-camel-case = "^1.4.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.272"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
